myapp/requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
gunicorn==25.3.0
packaging==26.1
psycopg2-binary==2.9.12
sqlparse==0.5.5
